gridinterpreter.cpp changed from if tree to switch case

+++ src/entity/playerStats.cpp
@@ -2,7 +2,7 @@
 #include "passiveItem.hpp"
 
 bool PlayerStats::PushBackPassiveItems(std::unique_ptr<PassiveItem> item) {
-	if (passiveItems.size() < inventorySize) {
+	if (passiveItems.size() < static_cast<size_t>(inventorySize)) {
 		passiveItems.push_back(std::move(item));
 		return true;
 	}
